+++ internal/reddit/reddit.go
@@ -4,6 +4,8 @@ import (
 	"github.com/bwmarrin/discordgo"
 	"github.com/codemicro/lgballtDiscordBot/internal/config"
 	"github.com/codemicro/lgballtDiscordBot/internal/state"
+	"math/rand"
+	"time"
 )
 
 func Start(state *state.State) error {
@@ -14,7 +16,12 @@ func Start(state *state.State) error {
 	}
 
 	for _, sub := range config.RedditFeeds {
-		go subMonitorSequencer(state, sub, ts)
+		go func() {
+			// Stagger timer starts - Reddit only likes you to send at most 1 request every two seconds.
+			// This is hacky way to try and ensure that happens
+			time.Sleep(time.Second * time.Duration(rand.Intn(28) + 2))
+			go subMonitorSequencer(state, sub, ts)
+		}()
 	}
 
 	return nil